[fix] move theme registration to src
