+++ .storybook-css/config.js
@@ -2,7 +2,7 @@ import { configure } from '@kadira/storybook';
 import ThemedStyleSheet from 'react-with-styles/lib/ThemedStyleSheet';
 import aphroditeInterface from 'react-with-styles-interface-aphrodite';
 import '../css/rheostat.css';
-import registerCSSInterfaceWithDefaultTheme from '../scripts/utils/registerCSSInterfaceWithDefaultTheme';
+import registerCSSInterfaceWithDefaultTheme from '../lib/utils/registerCSSInterfaceWithDefaultTheme';
 
 
 /* Register react with styles interface */
