+++ .storybook/config.js
@@ -1,7 +1,7 @@
 import { configure } from '@kadira/storybook';
 import ThemedStyleSheet from 'react-with-styles/lib/ThemedStyleSheet';
 import aphroditeInterface from 'react-with-styles-interface-aphrodite';
-import registerInterfaceWithDefaultTheme from '../scripts/utils/registerInterfaceWithDefaultTheme';
+import registerInterfaceWithDefaultTheme from '../lib/utils/registerInterfaceWithDefaultTheme';
 
 /* Register react with styles interface */
 registerInterfaceWithDefaultTheme(aphroditeInterface);
