+++ src/utils/registerCSSInterfaceWithDefaultTheme.js
@@ -1,5 +1,3 @@
-/* This file should be removed in the next major version update of Rheostat */
-
 import CSSInterface from 'react-with-styles-interface-css';
 
 import registerInterfaceWithDefaultTheme from './registerInterfaceWithDefaultTheme';
