+++ src/utils/registerInterfaceWithDefaultTheme.js
@@ -1,7 +1,5 @@
-/* This file should be removed in the next major version update of Rheostat */
-
 import ThemedStyleSheet from 'react-with-styles/lib/ThemedStyleSheet';
-import DefaultTheme from '../../src/themes/DefaultTheme';
+import DefaultTheme from '../themes/DefaultTheme';
 
 export default function registerInterfaceWithDefaultTheme(reactWithStylesInterface) {
   ThemedStyleSheet.registerInterface(reactWithStylesInterface);
